+++ lib/modules/highlight.d.ts
@@ -1,2 +1,0 @@
-export default highlightInit;
-declare function highlightInit(): any;