+++ lib/modules/index.d.ts
@@ -1,8 +1,0 @@
-import highlightInit from "./highlight";
-import { ImageDrop } from "./imagePasteDrop";
-import ImageResize from "./imageResize";
-import { MagicUrl } from "./magic-url";
-import MarkdownShortcuts from "./markdown-shortcuts";
-import ToolbarTable from "./toolbar-table";
-import QuillBetterTable from "./quill-better-table/quill-better-table";
-export { highlightInit, ImageDrop, ImageResize, MagicUrl, MarkdownShortcuts, ToolbarTable, QuillBetterTable };