+++ lib/modules/magic-url.d.ts
@@ -1,1 +1,0 @@
-export { default as MagicUrl } from "quill-magic-url";