+++ lib/modules/quill-better-table/utils/node-matchers.d.ts
@@ -1,3 +1,0 @@
-export function matchTableCell(node: any, delta: any, scroll: any): any;
-export function matchTableHeader(node: any, delta: any, scroll: any): any;
-export function matchTable(node: any, delta: any, scroll: any): any;