+++ lib/modules/toolbarHandler.d.ts
@@ -1,5 +1,0 @@
-export function imageUpload(quill: any, imgUploadApi: any, uploadSuccCB: any, uploadFailCB: any): void;
-export function linkHandler(value: any, ...args: any[]): void;
-export function undoHandler(): void;
-export function redoHandler(): void;
-export function uploadImg(blob: any, imgUploadApi: any, uploadSuccCB: any, uploadFailCB: any): void;