+++ lib/utils.d.ts
@@ -1,9 +1,7 @@
 import Quill from 'quill';
+import Delta from 'quill-delta';
 export declare function isUrl(url: string): boolean;
 export declare function isEmail(url: string): boolean;
-export declare function saveLink(quill: Quill & {
-    theme?: Record<string, any>;
-}, isToolbar: any): void;
 export declare function isMobile(): boolean;
-export declare function setContent(content: any, quill: Quill): void;
+export declare function setContent(content: Delta | string, quill: Quill): void;
 export declare const optionDisableToggle: (quill: Quill, blockList: string[], disable: boolean) => void;