+++ modules/imageResize.js
@@ -315,7 +315,7 @@ class Toolbar extends BaseModule {
           FloatStyle.add(this.img, 'left');
           MarginStyle.add(this.img, '0 1em 1em 0');
         },
-        isApplied: () => FloatStyle.value(this.img) === 'left'
+        isApplied: () => FloatStyle.value(this.img) === 'left',
       },
       {
         icon: IconAlignCenter,
@@ -324,7 +324,7 @@ class Toolbar extends BaseModule {
           FloatStyle.remove(this.img);
           MarginStyle.add(this.img, 'auto');
         },
-        isApplied: () => MarginStyle.value(this.img) === 'auto'
+        isApplied: () => MarginStyle.value(this.img) === 'auto',
       },
       {
         icon: IconAlignRight,
@@ -333,8 +333,8 @@ class Toolbar extends BaseModule {
           FloatStyle.add(this.img, 'right');
           MarginStyle.add(this.img, '0 0 1em 1em');
         },
-        isApplied: () => FloatStyle.value(this.img) === 'right'
-      }
+        isApplied: () => FloatStyle.value(this.img) === 'right',
+      },
     ];
   };